Message receiver now follows singleton pattern
Voices only get converted to text once on the server instead of once per client
