+++ YoutubeClient/Startup.cs
@@ -10,6 +10,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using YoutubeClient.Hubs;
+using ChatBot.MessageSpeaker;
+using System.IO;
+using Microsoft.AspNetCore.SignalR;
+using TwitchLib.Client.Events;
 
 namespace YoutubeClient
 {
@@ -19,6 +23,9 @@ namespace YoutubeClient
         /// integrate message receiver in a way thats accessible across the whole application
         /// </summary>
         public IConfiguration Configuration { get; }
+        private MessageReceiver messageReceiver = new MessageReceiver();
+        private GoogleTTSSettings googleTTSSettings = new GoogleTTSSettings();
+        private IHubContext<ChatHub> chatHub;
 
         public Startup(IConfiguration configuration)
         {
@@ -33,8 +40,11 @@ namespace YoutubeClient
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IHubContext<ChatHub> chatHub)
         {
+            messageReceiver.twitchClient.OnMessageReceived += Client_OnMessageReceived;
+            this.chatHub = chatHub;
+
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
@@ -60,6 +70,26 @@ namespace YoutubeClient
                     name: "default",
                     pattern: "{controller=Home}/{action=Index}/{id?}");
             });
+        }
+        private void Client_OnMessageReceived(object sender, OnMessageReceivedArgs e)
+        {
+            GetMessageAudio(e.ChatMessage.Username, e.ChatMessage.Message);
+        }
+
+        /// <summary>
+        /// Plays audio in the browser when recieving a message
+        /// </summary>
+        /// <returns></returns>
+        public void GetMessageAudio(string username, string message)
+        {
+            MemoryStream audioMemoryStream = new MemoryStream(GoogleTTSSettings
+                .GetVoiceAudio(username, message, googleTTSSettings.ttsClient).ToArray());
+
+            long audioLength = audioMemoryStream.Length;
+
+            chatHub.Clients.All.SendAsync("ReceiveMessageAudio", System.Convert.ToBase64String(audioMemoryStream.ToArray()));
+
+            return;
         }
     }
 }
